Update README.md
As mentioned in the discussion (https://github.com/dalenguyen/firestore-backup-restore/discussions/73) this code worked for me to get the sample working.
First time proposing a change. I hope this is the correct way to do it.

I also changed the initializeApp part in line 116. I think it needs to be done there as well, but I didn´t test that 'For local JSON' part of the code samples.